+++ src/app/about/components/Article.tsx
