+++ src/app/about/components/CaptionedImage.tsx
