+++ src/app/about/utils.ts
