+++ src/components/Slider.tsx
@@ -1,3 +1,4 @@
+'use client'
 import * as RadixSlider from '@radix-ui/react-slider'
 import clsx from 'clsx'
