+++ src/components/Toggle.tsx
@@ -1,3 +1,4 @@
+'use client'
 import clsx from 'clsx'
 import { useCallback, useState } from 'react'
