+++ src/components/Tooltip.tsx
@@ -1,3 +1,4 @@
+'use client'
 import React, { PropsWithChildren } from 'react'
 import * as TooltipPrimitive from '@radix-ui/react-tooltip'
