+++ src/features/pages/index.ts
@@ -1,7 +1,4 @@
 export { default as Freeplay } from './Freeplay'
-export { default as Home } from './Home'
-export { default as SelectSong } from './SelectSong'
-export { default as About } from './About'
 export { default as Training } from './Training'
 export { default as Speed } from './Training/Speed'
 export { PlaySong } from './PlaySong'
