+++ src/pages/_app.tsx
@@ -1,19 +1,13 @@
-import '@/styles/global.css'
+'use client'
 
 import type { AppProps } from 'next/app'
 import * as React from 'react'
 import { useRouter } from 'next/router'
 import Head from 'next/head'
-import { Inter } from 'next/font/google'
 
 import * as analytics from '@/features/analytics'
 import { TooltipProvider } from '@radix-ui/react-tooltip'
 import { addMetadata } from '@/features/data/library'
-
-const inter = Inter({
-  subsets: ['latin'],
-  variable: '--primary-font-family',
-})
 
 function App({ Component, pageProps }: AppProps) {
   const router = useRouter()
