+++ src/pages/_document.tsx
@@ -13,25 +13,6 @@ class MyDocument extends Document {
     return (
       <Html>
         <Head>
-          <meta charSet="utf-8" />
-
-          <meta name="author" content="Jake Fried" />
-          <meta name="description" content="🎹 Learn to play piano with sightread" />
-
-          <link rel="icon" href="/favicon.ico" />
-          <link rel="apple-touch-icon" href="/favicon.ico" />
-
-          <meta property="og:title" content="Sightread" />
-          <meta property="og:description" content="🎹 Free app for learning to play Piano" />
-          <meta property="og:image" content="images/mode_falling_notes_screenshot.png" />
-          <meta property="og:url" content="https://sightread.dev" />
-          <meta property="og:site_name" content="Sightread" />
-          <meta name="twitter:card" content="images/mode_falling_notes_screenshot.png" />
-          <meta
-            name="twitter:image:alt"
-            content="Sightread demo displaying falling notes visualization"
-          />
-
           {/* Global Site Tag (gtag.js) - Google Analytics */}
           <script async src={`https://www.googletagmanager.com/gtag/js?id=${GA_TRACKING_ID}`} />
           <script
