--- src/pages/about.tsx
@@ -1,3 +1,0 @@
-import { About } from '@/features/pages'
-
-export default About
